backend-python-rag-implementation/pyproject.toml
[project]
name = "backend-python-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "sentence-transformers>=5.3.0",
]
